+++ examples/rtc.rs
@@ -9,6 +9,7 @@ use cortex_m_rt::entry;
 use panic_halt as _;
 use rtt_target::{rprintln, rtt_init_print};
 
+use fugit::ExtU32;
 use stm32f4xx_hal::{pac, prelude::*, rtc::Rtc};
 use time::{
     macros::{date, time},