+++ examples/serial.rs
@@ -4,6 +4,7 @@
 use panic_halt as _;
 
 use cortex_m_rt::entry;
+use fugit::ExtU32;
 use stm32f4xx_hal as hal;
 
 use crate::hal::{pac, prelude::*};