+++ examples/stopwatch-with-ssd1306-and-interrupts-and-dma-i2c.rs
@@ -48,6 +48,7 @@ use embedded_graphics::{
     prelude::*,
     text::Text,
 };
+use fugit::ExtU32;
 use heapless::String;
 use ssd1306::{prelude::*, Ssd1306};