+++ examples/stopwatch-with-ssd1306-and-interrupts.rs
@@ -43,6 +43,7 @@ use embedded_graphics::{
     prelude::*,
     text::Text,
 };
+use fugit::ExtU32;
 use heapless::String;
 use ssd1306::{prelude::*, I2CDisplayInterface, Ssd1306};